RUB-259 fit utxo split under Codacy budget
